# Repository.openapi+json.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Repository API",
        "description": "",
        "version": "1.0"
    },
    "paths": {
...
        "/v1/demos/{demoId}": {
            "get": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "demoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "demoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/demos": {
            "get": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "gameTypes",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filterString",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "skipEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "takeEntries",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/DemoOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Demos"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/demos/{demoId}/file": {
            "post": {
                "tags": [
                    "Demos"
                ],
                "parameters": [
                    {
                        "name": "demoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DemoOrder": {
                "enum": [
                    "GameTypeAsc",
                    "GameTypeDesc",
                    "TitleAsc",
                    "TitleDesc",
                    "CreatedAsc",
                    "CreatedDesc",
                    "UploadedByAsc",
                    "UploadedByDesc"
                ],
                "type": "string",
                "format": "int32"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}